Moved model monitoring utilities to a dedicated directory
