+++ mlrun/utils/helpers.py
@@ -18,7 +18,6 @@ import json
 import re
 import sys
 import time
-from dataclasses import dataclass
 from datetime import datetime, timezone
 from importlib import import_module
 from os import environ, path
@@ -73,87 +72,6 @@ class StorePrefix:
             cls.FeatureSet,
             cls.FeatureVector,
         ]
-
-
-@dataclass
-class FunctionURI:
-    project: str
-    function: str
-    tag: Optional[str] = None
-    hash_key: Optional[str] = None
-
-    @classmethod
-    def from_string(cls, function_uri):
-        project, uri, tag, hash_key = parse_versioned_object_uri(function_uri)
-        return cls(
-            project=project, function=uri, tag=tag or None, hash_key=hash_key or None,
-        )
-
-
-@dataclass
-class VersionedModel:
-    model: str
-    version: Optional[str]
-
-    @classmethod
-    def from_string(cls, model):
-        try:
-            model, version = model.split(":")
-        except ValueError:
-            model, version = model, None
-
-        return cls(model, version)
-
-
-@dataclass
-class EndpointUID:
-    project: str
-    function: str
-    function_tag: str
-    function_hash_key: str
-    model: str
-    model_version: str
-    uid: Optional[str] = None
-
-    def __post_init__(self):
-        function_ref = (
-            f"{self.function}_{self.function_tag or self.function_hash_key or 'N/A'}"
-        )
-        versioned_model = f"{self.model}_{self.model_version or 'N/A'}"
-        unique_string = f"{self.project}_{function_ref}_{versioned_model}"
-        self.uid = hashlib.sha1(unique_string.encode("utf-8")).hexdigest()
-
-    def __str__(self):
-        return self.uid
-
-
-def create_model_endpoint_id(function_uri: str, versioned_model: str):
-    function_uri = FunctionURI.from_string(function_uri)
-    versioned_model = VersionedModel.from_string(versioned_model)
-
-    if (
-        not function_uri.project
-        or not function_uri.function
-        or not versioned_model.model
-    ):
-        raise ValueError("Both function_uri and versioned_model have to be initialized")
-
-    uid = EndpointUID(
-        function_uri.project,
-        function_uri.function,
-        function_uri.tag,
-        function_uri.hash_key,
-        versioned_model.model,
-        versioned_model.version,
-    )
-
-    return uid
-
-
-def parse_model_endpoint_store_prefix(store_prefix: str):
-    scheme, path = store_prefix.split("///", 1)
-    container, path = path.split("/", 1)
-    return scheme, container, path
 
 
 def get_artifact_target(item: dict, project=None):
